requirements.txt
aniso8601==10.0.1
apispec==6.8.2
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
diskcache==5.6.3
dnspython==2.7.0
exceptiongroup==1.3.0
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
flask-smorest==0.46.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
mongoengine==0.29.1
numpy==2.2.6
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pytest==8.4.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
webargs==8.7.0
Werkzeug==3.1.3
